+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/ui/fragments/FirstFragment.kt
@@ -89,7 +89,7 @@ class FirstFragment : Fragment() {
                                 is NavigationAction.NavigateToWebView -> {
                                     Logger.logScreenView("web_view_fragment")
                                     val bundle = Bundle().apply {
-                                        putString(WebViewFragment.ARG_URL, action.url)
+                                        putString(WebViewFragment.ARG_DEMO_KEY, action.demoKey)
                                     }
                                     findNavController().navigate(R.id.action_FirstFragment_to_WebViewFragment, bundle)
                                 }